cambios

+++ src/main/java/org/esfe/repository/ClienteRepository.java
@@ -3,7 +3,7 @@ package org.esfe.repository;
 import org.esfe.entities.Cliente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClienteRepository extends JpaRepository<Cliente, Long> {
+public interface ClienteRepository extends JpaRepository<Cliente, Integer> {
     Cliente findByCorreo(String correo);
     Cliente findByDui(String dui);
 }
